+++ src/components/IconButton/IconButton.js
@@ -1,0 +1,11 @@
+import './IconButton.css';
+
+const IconButton = ({ CustomIcon, name }) => {
+	return (
+		<div className='IconButton-btn'>
+			{CustomIcon} {name}
+		</div>
+	);
+};
+
+export default IconButton;
