--- src/pages/ProductsEditPage/components/ProductsEditBtn/ProductsEditBtn.js
@@ -1,12 +1,0 @@
-import './ProductsEditBtn.css';
-
-
-const ProductsEditBtn = ({ CustomIcon, name }) => {
-	return (
-		<div className='productsEditPage-btn'>
-			{CustomIcon} {name}
-		</div>
-	);
-};
-
-export default ProductsEditBtn;
